Change mkdir() / rmdir() / unlink() raise warning behavior.
According to PHP5, mkdir() / rmdir() / unlink() raises warning only
if dealing with plain files.

Fixed vfsstream framework test regressions.

+++ hphp/runtime/ext/std/ext_std_file.cpp
@@ -1520,15 +1520,20 @@ bool HHVM_FUNCTION(unlink,
   CHECK_PATH_FALSE(filename, 1);
   Stream::Wrapper* w = Stream::getWrapperFromURI(filename);
   if (!w) return false;
-  if (w->unlink(filename) != 0) {
-    raise_warning(
-      "%s(%s): %s",
-       __FUNCTION__ + 2,
-       filename.c_str(),
-       folly::errnoStr(errno).c_str()
-    );
-    return false;
+  auto fsw = dynamic_cast<FileStreamWrapper*>(w);
+  if (fsw != nullptr) {
+    if (w->unlink(filename) != 0) {
+      raise_warning(
+        "%s(%s): %s",
+         __FUNCTION__ + 2,
+         filename.c_str(),
+         folly::errnoStr(errno).c_str()
+      );
+      return false;
+    }
+    return true;
   }
+  CHECK_SYSTEM_SILENT(w->unlink(filename));
   return true;
 }
 
@@ -1731,7 +1736,12 @@ bool HHVM_FUNCTION(mkdir,
   Stream::Wrapper* w = Stream::getWrapperFromURI(pathname);
   if (!w) return false;
   int options = recursive ? k_STREAM_MKDIR_RECURSIVE : 0;
-  CHECK_SYSTEM(w->mkdir(pathname, mode, options));
+  auto fsw = dynamic_cast<FileStreamWrapper*>(w);
+  if (fsw != nullptr) {
+    CHECK_SYSTEM(fsw->mkdir(pathname, mode, options));
+    return true;
+  }
+  CHECK_SYSTEM_SILENT(w->mkdir(pathname, mode, options));
   return true;
 }
 
@@ -1741,7 +1751,12 @@ bool HHVM_FUNCTION(rmdir,
   Stream::Wrapper* w = Stream::getWrapperFromURI(dirname);
   if (!w) return false;
   int options = 0;
-  CHECK_SYSTEM(w->rmdir(dirname, options));
+  auto fsw = dynamic_cast<FileStreamWrapper*>(w);
+  if (fsw != nullptr) {
+    CHECK_SYSTEM(w->rmdir(dirname, options));
+    return true;
+  }
+  CHECK_SYSTEM_SILENT(w->rmdir(dirname, options));
   return true;
 }
